Todo

Description

To-do lists can be represented in three different ways:

String: a simple line-separated representation for saving lists to disk;

List: an internal representation;

JSON: a JSON representation for passing list data to the front-end.

Disk Representation: Each line represents a to-do, and begins either with - to indicate an incomplete item, or x to indicate a completed one. Note that the list is ordered. Eg:

 - One
 x Two
 - Three

Internal Representation: A List is a list of Items. Each item is either a `Complete itemBody` or an `Incomplete itemBody`. Eg:

 [ Complete "One", Incomplete "Two", Complete "Three" ]

JSON Representation: Items in the JSON representation are explicity numbered. This corresponds to the numbering implicit in the ordering of the other representations. This is so a single item may be modified by the front-end without having to resend the entire list. Eg:

 [
   { body: "one",   done: false },
   { body: "two",   done: true  },
   { body: "three", done: false }
 ]

This module exports functions for manipulating the lists stored in listDirectory. There's not a lot of interesting logic - just conversion between the different representations, and a bunch of file IO.

Documentation

loadList :: String -> IO List

Read the contents of listDirectory/listName, parse it, and return a List

addItem :: String -> Item -> IO ()

Append item to listDirectory/listName

editItem :: String -> Int -> Item -> IO ()

Replace the itemIdth item in listDirectory/listName with item

removeItem :: String -> Int -> IO ()

Remove the itemIdth item from listDirectory/listName

showLists :: IO [String]

Return the names of the files in listDirectory.

data Item

Constructors

Complete

Fields

itemBody :: String

Incomplete

Fields

itemBody :: String

data List

Constructors

List [Item]